refactor: AddProductPanel 추가
