+++ src/refactoring/components/admin/ProductManagement/AddProductPanel/NewProductForm.tsx
