+++ src/refactoring/components/admin/ProductManagement/AddProductPanel/index.ts
@@ -1,0 +1,1 @@
+export { AddProductPanel } from './AddProductPanel';
