+++ src/refactoring/components/admin/ProductManagement/ProductAccordion/index.ts
@@ -1,0 +1,2 @@
+export { ProductAccordion } from './ProductAccordion';
+export { useToggleProductAccordion } from './useToggleProductAccordion';
